+++ includes/requirements.php
@@ -1,6 +1,6 @@
 <?php
 /*
-	Copyright 2015 Axelerant (email: info@axelerant.com)
+	Copyright 2015 Axelerant
 
 	This program is free software; you can redistribute it and/or modify
 	it under the terms of the GNU General Public License, version 2, as