+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 1.6.1RC1
 * Require Aihrus Framework 1.2.2
+* Update store branding
 
 ## 1.6.0
 * Add FAQ maximum bulk edit limit